refactor: separate revert test suite into its own file
